+++ src/main/java/eu/olli/cowmoonication/listener/PlayerListener.java
@@ -21,5 +21,6 @@ public class PlayerListener {
     @SubscribeEvent
     public void onServerLeave(FMLNetworkEvent.ClientDisconnectionFromServerEvent e) {
         main.getFriends().saveBestFriends();
+        main.getPlayerCache().clearAllCaches();
     }
 }