+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## 1.8.9-0.5.0 - unreleased
+## [1.8.9-0.5.0] - unreleased
+### Added
+- Added Tab-completable usernames for several commands (e.g. party, msg, boop, ...)
+  - use command `/tabcompletablecommand` for a full list of commands that support Tab-completion
+  - the list of supported commands can be modified via `/moo config`
+  - Tab-completable usernames consist of: the last 50 players that typed in private/friends, Guild or Party chat; up to 50 best friends that are currently logged in
+
 ### Changed
 - Improved player stalking feature:
   - now includes 'offline for &lt;duration&gt;'
   - better handling of special cases (e.g. nicked players, players who haven't joined in years)
+- Various Code refactorings 
 
 ## [1.8.9-0.4.0] - 18.04.2020
 ### Added
@@ -59,6 +66,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 *Note:* The 'best friends' list is currently available via <kbd>ESC</kbd> > Mod Options > Cowmoonication > Config > bestFriends.
 
+[1.8.9-0.5.0]: https://github.com/cow-mc/Cowmoonication/compare/v1.8.9-0.4.0...master
 [1.8.9-0.4.0]: https://github.com/cow-mc/Cowmoonication/compare/v1.8.9-0.3.1...v1.8.9-0.4.0
 [1.8.9-0.3.1]: https://github.com/cow-mc/Cowmoonication/compare/v1.8.9-0.3.0...v1.8.9-0.3.1
 [1.8.9-0.3.0]: https://github.com/cow-mc/Cowmoonication/compare/v1.8.9-0.2.0...v1.8.9-0.3.0